separate  middle banners
